Added Dynamic allocator

+++ engine/src/renderer/vulkan/shaders/material/vulkan_material_shader.c
@@ -105,7 +105,7 @@ b8 vulkan_material_shader_create(vulkan_context* context, vulkan_material_shader
     scissor.extent.height = context->framebuffer_height;
 
     // Attributes
-    u32 offset = 0;
+    u64 offset = 0;
 #define ATTRIBUTE_COUNT 2
     VkVertexInputAttributeDescription attribute_descriptions[ATTRIBUTE_COUNT];
     // Position, texcoord
