+++ tests/src/containers/freelist_tests/freelist_tests.c
@@ -47,7 +47,7 @@ u8 freelist_should_allocate_one_and_free_one() {
     freelist_create(total_size, &memory_requirement, block, &list);
 
     // Allocate some space.
-    u32 offset = INVALID_ID;  // Start with invalid id, which is a good default since it should never happen.
+    u64 offset = INVALID_ID;  // Start with invalid id, which is a good default since it should never happen.
     b8 result = freelist_allocate_block(&list, 64, &offset);
     // Verify that result is true, offset should be set to 0.
     expect_to_be_true(result);
@@ -87,21 +87,21 @@ u8 freelist_should_allocate_one_and_free_multi() {
     freelist_create(total_size, &memory_requirement, block, &list);
 
     // Allocate some space.
-    u32 offset = INVALID_ID;  // Start with invalid id, which is a good default since it should never happen.
+    u64 offset = INVALID_ID;  // Start with invalid id, which is a good default since it should never happen.
     b8 result = freelist_allocate_block(&list, 64, &offset);
     // Verify that result is true, offset should be set to 0.
     expect_to_be_true(result);
     expect_should_be(0, offset);
 
     // Allocate some more space.
-    u32 offset2 = INVALID_ID;  // Start with invalid id, which is a good default since it should never happen.
+    u64 offset2 = INVALID_ID;  // Start with invalid id, which is a good default since it should never happen.
     result = freelist_allocate_block(&list, 64, &offset2);
     // Verify that result is true, offset should be set to the size of the previous allocation.
     expect_to_be_true(result);
     expect_should_be(64, offset2);
 
     // Allocate one more space.
-    u32 offset3 = INVALID_ID;  // Start with invalid id, which is a good default since it should never happen.
+    u64 offset3 = INVALID_ID;  // Start with invalid id, which is a good default since it should never happen.
     result = freelist_allocate_block(&list, 64, &offset3);
     // Verify that result is true, offset should be set to the offset+size of the previous allocation.
     expect_to_be_true(result);
@@ -121,7 +121,7 @@ u8 freelist_should_allocate_one_and_free_multi() {
     expect_should_be(total_size - 128, free_space);
 
     // Allocate some more space, this should fill the middle block back in.
-    u32 offset4 = INVALID_ID;  // Start with invalid id, which is a good default since it should never happen.
+    u64 offset4 = INVALID_ID;  // Start with invalid id, which is a good default since it should never happen.
     result = freelist_allocate_block(&list, 64, &offset4);
     // Verify that result is true, offset should be set to the size of the previous allocation.
     expect_to_be_true(result);
@@ -170,21 +170,21 @@ u8 freelist_should_allocate_one_and_free_multi_varying_sizes() {
     freelist_create(total_size, &memory_requirement, block, &list);
 
     // Allocate some space.
-    u32 offset = INVALID_ID;  // Start with invalid id, which is a good default since it should never happen.
+    u64 offset = INVALID_ID;  // Start with invalid id, which is a good default since it should never happen.
     b8 result = freelist_allocate_block(&list, 64, &offset);
     // Verify that result is true, offset should be set to 0.
     expect_to_be_true(result);
     expect_should_be(0, offset);
 
     // Allocate some more space.
-    u32 offset2 = INVALID_ID;  // Start with invalid id, which is a good default since it should never happen.
+    u64 offset2 = INVALID_ID;  // Start with invalid id, which is a good default since it should never happen.
     result = freelist_allocate_block(&list, 32, &offset2);
     // Verify that result is true, offset should be set to the size of the previous allocation.
     expect_to_be_true(result);
     expect_should_be(64, offset2);
 
     // Allocate one more space.
-    u32 offset3 = INVALID_ID;  // Start with invalid id, which is a good default since it should never happen.
+    u64 offset3 = INVALID_ID;  // Start with invalid id, which is a good default since it should never happen.
     result = freelist_allocate_block(&list, 64, &offset3);
     // Verify that result is true, offset should be set to the offset+size of the previous allocation.
     expect_to_be_true(result);
@@ -205,7 +205,7 @@ u8 freelist_should_allocate_one_and_free_multi_varying_sizes() {
 
     // Allocate some more space, this time larger than the old middle block. This should have a new offset
     // at the end of the list.
-    u32 offset4 = INVALID_ID;  // Start with invalid id, which is a good default since it should never happen.
+    u64 offset4 = INVALID_ID;  // Start with invalid id, which is a good default since it should never happen.
     result = freelist_allocate_block(&list, 64, &offset4);
     // Verify that result is true, offset should be set to the size of the previous allocation.
     expect_to_be_true(result);
@@ -254,7 +254,7 @@ u8 freelist_should_allocate_to_full_and_fail_to_allocate_more() {
     freelist_create(total_size, &memory_requirement, block, &list);
 
     // Allocate all space.
-    u32 offset = INVALID_ID;  // Start with invalid id, which is a good default since it should never happen.
+    u64 offset = INVALID_ID;  // Start with invalid id, which is a good default since it should never happen.
     b8 result = freelist_allocate_block(&list, 512, &offset);
     // Verify that result is true, offset should be set to 0.
     expect_to_be_true(result);
